pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pytest-fixedpoint"
version = "0.1.0"
description = "Pytest plugin for recording and replaying deterministic function calls"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "pytest>=7.0",
    "pyyaml>=6.0",
]
classifiers = [
    "Framework :: Pytest",
    "Programming Language :: Python :: 3",
]

[project.entry-points.pytest11]
fixedpoint = "fixedpoint.plugin"

[tool.hatch.build.targets.wheel]
packages = ["src/fixedpoint"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]

[tool.ruff.lint.isort]
known-first-party = ["fixedpoint"]
